Discovery/requirements.txt
Appium_Python_Client==2.6.1
pytest==7.0.1
requests==2.27.1
selenium==4.4.3
urllib3==1.26.8
